Call `pip` directly instead of `python -m pip`.
The former is required to make our scripts compatible both with and without
Docker (alternative user installation scheme vs. virtualenv).

The latter is required to work-around issues with spaces in shebangs for
scripts in the virtualenv bin directory, and can be resolved by not using
a directory with a space in it or with an alias.