Source local dotenv file first, then encrypted dotenv file (if specified).
The local dotenv file is not encrypted or version controlled, and may
contain the password needed to decrypt secret files.

Previously, we only sourced the local dotenv file if no encrypted dotenv
file was specified. Now we source both.

An encrypted dotenv file will override the local dotenv file, which
shouldn't normally be a problem, because we previously only used the
local dotenv file as a fallback, and when we want to source an encrypted
dotenv file, usually no local dotenv file exists.