Add redis-cache.py script to cache & check values across hosts
The redis-cache.py script exposes simple actions
to set and get values in Redis, but more importantly
includes the 'set-and-match' action that will:

 - check for a key value in the Redis cache
 - check whether it matches a provided input value
 - return an ERRORLEVEL of 0 if the cached and
   input values match, otherwise a 100 value
 - in either case, set/update the cached value in
   Redis.

The 'set-and-match' action should be useful in
combination with scripts run with waitlock to
cache values and check across hosts/containers to
determine whether tasks such as DB migrations
should be run.

Commit includes shell script tests that also
serve as example usage of set-and-match.