Run dev server on `NGINX_PORT` (default: 8000).
When running multiple processes without isolation, we need to run them
on different ports.

We use `NGINX_PORT` because the dev server is typically used instead of
(not behind) nginx, which may in turn behind HAproxy. This allows us to
quickly `supervisor.sh stop nginx` and `runserver.sh` to run the server
from an interactive process.