Wrap supervisor programs with `entrypoint.sh`.
Previously, even though supervisor itself is wrapped with `entrypoint.sh`,
and passes on its environment to child processes, when a dotenv file is
updated we need to restart supervisor itself to pickup the changes instead
of just restarting the affected program.

Now, we execute `entrypoint.sh` once for supervisor and again for every
program and every restart, but `entrypoint.sh` should be quick to execute
and not do anything besides configure the environment.