Disable `call_command` task and add new `clearsessions` task.
The `call_command` task was a bad idea because we need to use concurrency
decorator to either enqueue, fail or skip when a task is already running.

By its general nature, multiple `call_command` tasks could be scheduled
to run at the same time. We'd have to avoid these concurrency decorators,
and then we may get multiple concurrent executions.