Use a shim to execute `python` (default) or `python3` based on `PYTHON_VERSION`.
We need this because shebang lines must be an absolute path but `python`
will always be a symlink to `python2`, and replacing it with a symlink to
`python3` breaks Supervisor which is not available for Python 3.

We can't use a virtualenv (which could symlink `python` to `python3`) inside
Docker images because we need to take advantage of the "user" alternate
installation scheme to avoid reinstalling the entire virtualenv during local
development, even when there are no changes.

Python shim.

Typo.