Only install `entrypoint.sh` into virtualenv or system `bin` directory.
This should make it a little easier to work find and work on the right file
during development by avoiding duplication in the project tree.

Ideally we wouldn't even need to install this, but then `go.sh` would need
to be executed and discover the `ixc_django_docker/bin` directory when running
both with and without Docker.

Maybe in a future update.