Do not hard code database scheduler for celery beat.
This can still be enabled in project settings, if needed. But we should
avoid it if we can because it makes it difficult to disable tasks based
on the runtime environment.

For example, we often sync production database to staging for testing,
but we may not want to run all the same periodic tasks on staging.