Downgrade Dockerize so we can use `-stdout` on log files that do not yet exist.
In 0.6 Dockerize can monitor named pipes, but a side effect is that monitored files must exist at the time dockerize is executed.

This is a problem when we wrap the command (e.g. supervisor) that produces the monitored log file for the first time.

An ugly workaround is to touch the log file in the entrypoint before calling dockerize.