Don't use redis-cache.py anymore, now that we have shared persistent volumes.
Revert back to storing checksums on the local file system to avoid having
to wait for Redis on startup, and to avoid potentially dangerous failures
if the checksum is cleared from Redis for whatever reason.

For example, previously if the `redis` service was restarted for any reason,
the `setup-git-commit` hash might be lost. Then if production gunicorn or
other services were also restarted, they would wait forever for setup to
run again, even though it had already run for that hash.