Add whitenoise middleware in the right order
The only reason why things kept working without this change is because
the sites it was used on had a catch-all URL pattern (usually provided
by Wagtail), which would provide a resolver match for staticfile
requests, preventing CommonMiddleware from appending a slash and
redirecting.

The whitenoise documentation specifially instructs to place it at the
top, excepting Security Middleware, and this change may actually result
in a slight performance boost.