Expect `manage.py` to live in the project root again, and not in the `ixc_django_docker/bin` directory.
We previously moved this into `ixc-django-docker` for the convenience of being able to execute it from any directory, and to activate `ixc-django-docker` settings.

But Django developers expect an executable `manage.py` file in the project root.

To activate `ixc-django-docker` settings, we now export `DJANGO_SETTINGS_MODULE` in `entrypoint.sh`, but ONLY if it is not already set. Individual projects can export a different value in their `Dockerfile` / `Dockerfile.sh` or `manage.py` files, if they want to use another settings module.