Codefresh pushes untested Docker image to make it cacheable
Push the built but pre-test Docker image in parallel with the tests
run so that the image will be available for fetching from Docker Hub
if necessary, e.g. if you re-run the Codefresh run, or want to fetch
and debug a failing image locally.

This is mainly motivated by the fact that Codefresh no longer
provides its own repository for images, which it used to, and thus
no longer caches every built image including those that failed the
`run_tests` step.