Make DOTENV available as a default setting
DOTENV is useful in many places, such as setting unique names for
things like databases or ElasticSearch indexes etc.

Make it easily available as a setting that is guaranteed to be
present, and given a clear default value if not, instead of requiring
projects to hunt for and handle the 'DOTENV' environment variable.