Update go.sh
Do not install required python version automatically. Instead, abort with an error so the user can install it manually. This fixes a faulty assumption that the newly installed version will actually be used, by again validating the installed version on the next run.

Always validate the virtualenv Python version, instead of only validating when the available Python version does not match the required version. We would have previously automatically installed the required version and assumed that it would be used to create the virtualenv.