Remove redundant `npm run build` step
This step is redundant in setup.sh because it gets run as part of
the recommended Dockerfile when building an image. Before this change
static files etc would get generated by the Codefresh build, then
again slowly and pointlessly when you run the setup service.

So either this command should be removed altogether from this script,
or maybe be adjusted to run only in a local development context?