Compile Sass into the source directory again.
This was disabled to avoid recursive compilation with `--watch`, when each compiled `.css` file triggered a recompile. But that was a bug in dart-sass, apparently fixed in 1.26.3.

We need to compile to the `static` directory instead of `static_root` so that compiled Sass files can be collected and compressed with a unique filename by the whitenoise storage backend.