Update abandoned ConcurrentLogHandler to the maintained concurrent-log-handler fork
This is prompted by the face that setuptools no longer supports the
`use_2to3` directive used by the old ConcurrentLogHandler.

The maintained concurrent-log-handler fork avoids the use of that
directive and claims to be backwards-compatible, though this hasn't been
tested.

If this change causes problems, one solution is to revert this change
and pin setuptools to version <58.

However, because `use_2to3` is scheduled for deprecation in Python 3.10,
this is not a future-proof approach, and it would be better to find a
way to make the fork work.

But if in a rush, pinning setuptools is a quick an effective way to stop
the bleeding.