Temporarily disable nose-progressive due to setuptools>=58 incompatibility
See previous commit for details.

Again, we could just pin setuptools to <58, but due to the upcoming
deprecation of `use_2to3`, it's more future-proof to tackle the issue
head-on.

Since nose-progressive is mainly decorative, it seems appropriate to
disable its usage for now, until we or someone else release a version of
nose-progressive with Python 3 support that does not rely on 2to3.